refactor: seperate mqtt, canbus, wifi and NTP methods into their own files
